requirements.txt
gamspy>=0.13.0
pandas>=2.0.0
pyyaml>=6.0
openpyxl>=3.1.0
